upng: create and tree-shake lib
